Created base class and updated gitignore
Created a base class for the quicksort implementations. This base class
defines comparison operations which also keep track of the number
of comparisons done in total. Similarly, the class also defines
a method which swaps two elements in the data and keeps track of the
total number of swaps.

Updated the .gitignore to ignore PyCharm files