Massive refactoring
doInsertionSort, pivotSelectionMethod and insertionSortThreshold have been
moved to BaseQuicksort. This definitely cleans up how data is collected.

collectData now also collects data for different pivot selection methods.
This added a bunch of code which was defintely redundant so I refactored
it to hopefully be cleaner. There are more methods which are more concise.
Docs need to be written though :(